Made sort parameter an array and added nullsfirst option Updated review sort calls Moved types and schema into another file and updated their imports
