+++ src/app/(main)/[productType]/[coverType]/[make]/[model]/[year]/page.tsx
@@ -1,9 +1,8 @@
-import { TReviewData, getProductData, getProductMetadata } from '@/lib/db';
+import { getProductData, getProductMetadata } from '@/lib/db';
 import { notFound, redirect } from 'next/navigation';
 import CarPDP from '@/app/(main)/[productType]/components/CarPDP';
 import { TPathParams } from '@/utils';
 import {
-  TProductReviewSummary,
   // filterReviewImages,
   getAllReviewsWithImages,
   getProductReviewSummary,
@@ -11,6 +10,7 @@ import {
 } from '@/lib/db/review';
 import { deslugify } from '@/lib/utils';
 import { PREMIUM_PLUS_URL_PARAM } from '@/lib/constants';
+import { TProductReviewSummary, TReviewData } from '@/lib/types/review';
 
 //TODO: Refactor code so we can generate our dynamic paths as static HTML for performance
 
@@ -100,6 +100,10 @@ export default async function CarPDPDataLayer({
               page: 0,
               limit: 8,
             },
+            sort: [
+              { field: 'sku', order: 'asc' },
+              { field: 'helpful', order: 'desc', nullsFirst: false },
+            ],
           }
         ),
         getProductReviewSummary({
