+++ src/app/(main)/[productType]/[coverType]/[make]/[model]/page.tsx
@@ -1,9 +1,7 @@
-import { TReviewData, getAllModels, getProductData } from '@/lib/db';
-import { notFound, redirect } from 'next/navigation';
+import { getAllModels, getProductData } from '@/lib/db';
+import { notFound } from 'next/navigation';
 import CarPDP from '@/app/(main)/[productType]/components/CarPDP';
 import {
-  TProductReviewSummary,
-  // filterReviewImages,
   getAllReviewsWithImages,
   getProductReviewSummary,
   getProductReviewsByPage,
@@ -11,6 +9,7 @@ import {
 import { TPathParams } from '@/utils';
 import { deslugify } from '@/lib/utils';
 import { PREMIUM_PLUS_URL_PARAM } from '@/lib/constants';
+import { TReviewData, TProductReviewSummary } from '@/lib/types/review';
 
 //TODO: Refactor code so we can generate our dynamic paths as static HTML for performance
 export const revalidate = 0;
@@ -83,6 +82,10 @@ export default async function CarPDPDataLayer({
               page: 0,
               limit: 8,
             },
+            sort: [
+              { field: 'sku', order: 'asc' },
+              { field: 'helpful', order: 'desc', nullsFirst: false },
+            ],
           }
         ),
         getProductReviewSummary({
