+++ src/app/(main)/[productType]/[coverType]/[make]/page.tsx
@@ -88,6 +88,10 @@ export default async function CarPDPDataLayer({
               page: 0,
               limit: 8,
             },
+            sort: [
+              { field: 'sku', order: 'asc' },
+              { field: 'helpful', order: 'desc', nullsFirst: false },
+            ],
           }
         ),
         getProductReviewSummary({
