+++ src/app/(main)/seat-covers/[coverType]/[make]/[model]/page.tsx
@@ -5,16 +5,15 @@ import {
   getSeatCoverProductsByDisplayColor,
 } from '@/lib/db/seat-covers';
 import {
-  TProductReviewSummary,
-  TReviewData,
   getAllReviewsWithImages,
   getProductReviewSummary,
   getProductReviewsByPage,
 } from '@/lib/db/review';
 import { deslugify } from '@/lib/utils';
 import SeatCoverDataWrapper from '../../../components/SeatCoverDataWrapper';
+import { TReviewData, TProductReviewSummary } from '@/lib/types/review';
 
-export const revalidate = 0
+export const revalidate = 0;
 
 export type TCarCoverSlugParams = {
   make: string;
@@ -83,6 +82,10 @@ export default async function SeatCoverDataLayer({
               page: 0,
               limit: 8,
             },
+            sort: [
+              { field: 'sku', order: 'asc' },
+              { field: 'helpful', order: 'desc', nullsFirst: false },
+            ],
           }
         ),
         getProductReviewSummary({
