+++ src/lib/db/index.ts
@@ -17,7 +17,6 @@ const supabaseKey = process.env.NEXT_PUBLIC_SUPABASE_KEY ?? '';
 const supabase = createClient<Database>(supabaseUrl, supabaseKey);
 
 export type TInitialProductDataDB = Tables<'Products'>;
-export type TReviewData = Tables<'reviews-2'>;
 
 //If the table you want to access isn't listed in TableRow,
 //generate new types in the Supabase dashboard to update
